+++ CHANGELOG.md
@@ -6,3 +6,12 @@
 # 2/8/2021
 - Added CSS to all components except marker.js
 - Fixed marker interaction with "Add Marker" mode
+
+# 2/9/2021
+- Added search bar UI and moved "Add Marker" button
+
+# 2/10/2021
+- Added Google Places package and UI for search bar
+- Added Google Maps styles for marker mode
+- Added marker showing for location search 
+